requirements.txt
streamlit==1.37.1
python-slugify==8.0.4
